Fix build --install failure in SDK top level
When running build.py --install from top level, do not add clashing packaging/install rules.
The original if condition here only applied when --package was selected,
not for --install.
What we want is just to check if we are running from component level,
and only in that case bother with the packaging rules.

Change-Id: I0ff919d4bca5bc1818c68c0faef2c2df2a80500e
Signed-off-by: Gergely Nagy <gergely.nagy@arm.com>